+++ pjt-backend/src/services/services.service.ts
@@ -1,4 +1,4 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import { Injectable, NotFoundException, BadRequestException } from '@nestjs/common';
 import { PrismaService } from '@/prisma/prisma.service';
 
 @Injectable()
@@ -29,6 +29,19 @@ export class ServicesService {
   }
 
   async remove(id: string) {
+    const service = await this.prisma.service.findUnique({ where: { id } });
+    if (!service) {
+      throw new NotFoundException('Serviço não encontrado');
+    }
+    
+    const appointmentServicesCount = await this.prisma.appointmentService.count({
+      where: { serviceId: id }
+    });
+    
+    if (appointmentServicesCount > 0) {
+      throw new BadRequestException('Não é possível excluir serviço com agendamentos');
+    }
+    
     await this.prisma.service.delete({ where: { id } });
   }
 }
